Updating checkbox in every table
