+++ src/components/Trends/TrendsTable/TrendsTables.js
@@ -1,10 +1,41 @@
 import React, { PureComponent } from 'react'
+import { connect } from 'react-redux'
+import { TRENDS_SELECTED_WORDS } from '../../../components/Trends/actions'
 import TrendsTable from './TrendsTable'
 import { dateDifferenceInWords, HOUR } from '../../../utils/dates'
 import styles from './TrendsTables.module.scss'
 
 class TrendsTables extends PureComponent {
+  static defaultProps = {
+    selectedTrends: new Set()
+  }
+
+  state = {
+    selected: this.props.selectedTrends
+  }
+
+  componentWillUnmount () {
+    const { selected } = this.state
+    if (selected.size > 0) {
+      this.props.setSelectedTrends(selected)
+    }
+  }
+
+  selectTrend = trend => {
+    const { selected: oldSelected } = this.state
+    const selected = new Set([...oldSelected])
+
+    if (selected.has(trend)) {
+      selected.delete(trend)
+    } else {
+      selected.add(trend)
+    }
+
+    this.setState({ selected })
+  }
+
   render () {
+    const { selected } = this.state
     const { trends, isLoading, selectable, isLoggedIn } = this.props
 
     return (
@@ -23,6 +54,8 @@ class TrendsTables extends PureComponent {
                 className={styles.table}
                 trend={trend}
                 isLoggedIn={isLoggedIn}
+                selectTrend={this.selectTrend}
+                selectedTrends={selected}
               />
             )
           })}
@@ -32,11 +65,29 @@ class TrendsTables extends PureComponent {
           trend={trends.length > 0 ? trends[trends.length - 1] : {}}
           header='Last trends'
           selectable={selectable}
-          isLoggedIn={isLoggedIn}
+          isLoggedIn={isLoggedIn || true}
+          selectTrend={this.selectTrend}
+          selectedTrends={selected}
         />
       </div>
     )
   }
 }
 
-export default TrendsTables
+const mapStateToProps = ({
+  hypedTrends: { selectedTrends },
+  user: { token }
+}) => ({
+  selectedTrends,
+  isLoggedIn: !!token
+})
+
+const mapDispatchToProps = dispatch => ({
+  setSelectedTrends: payload =>
+    dispatch({ type: TRENDS_SELECTED_WORDS, payload })
+})
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(TrendsTables)
